+++ src/store/utility.ts
@@ -1,4 +1,5 @@
 import { State } from "@/store/type";
+import { ToolbarButtonTagType } from "@/type/preload";
 
 export function sanitizeFileName(fileName: string): string {
   // \x00 - \x1f: ASCII 制御文字
@@ -43,3 +44,19 @@ export function buildProjectFileName(state: State, extension?: string): string {
     ? `${defaultFileNameStem}.${extension}`
     : defaultFileNameStem;
 }
+
+export const getToolbarButtonName = (tag: ToolbarButtonTagType): string => {
+  const tag2NameObj: Record<ToolbarButtonTagType, string> = {
+    PLAY_CONTINUOUSLY: "連続再生",
+    STOP: "停止",
+    EXPORT_AUDIO_ONE: "１つ書き出し",
+    EXPORT_AUDIO_ALL: "全部書き出し",
+    EXPORT_AUDIO_CONNECT_ALL: "音声を繋げて書き出し",
+    SAVE_PROJECT: "プロジェクト保存",
+    UNDO: "元に戻す",
+    REDO: "やり直す",
+    IMPORT_TEXT: "テキスト読み込み",
+    EMPTY: "空白",
+  };
+  return tag2NameObj[tag];
+};
